requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
streamlit>=1.30.0
pillow>=9.0.0